Bug 887701 - Ensure proper serialization of upgrade path. r=gene.lian, a=leo+
Most of IndexedDB is relying on asynchronicity. This is fun, but in the
present case, it can trigger bad behavior: upgrade being performed in
the wrong order. This can be a big issue if one further update depends
on a previous one, which is the case for adding messageRecord.type.